Resolved possible segmentation fault.
The ruter_line_copy() function used memcpy() to create a copy of its line_t structure, without creating a separate copy of the 'name' field. Anyone reading this field would risk accessing freed memory, which would cause a segmentation fault. The structure is now copied field by field in a way that ensures that any field not explicitly copied will be zero, or NULL (in the case of a pointer). The 'name' field is now correctly duplicated.