Renamed CurlClient to WebClient and added WebResponse.
The CurlClient class has been renamed to WebClient. The Request() method
of a WebClient object accepts an argument of type WebRequest which
specifies the properties of the connection. The Request() method will
eventually be modified to return an object of type WebResponse which
will describe all the details of the connection attempt after the fact.